% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{MultiPointsPlot}
\alias{MultiPointsPlot}
\title{Time series curves for two points}
\usage{
MultiPointsPlot(N, Id1, Id2, Id3, Id4, Id5)
}
\arguments{
\item{Id1}{-  ID number for point 1}

\item{Id2}{-  Id number for point 2}
}
\value{
Two time series curves together in a single plot (The first point represented as a soild line and the second point as a dotted line)
}
\description{
TwoPointsPlot function takes the ID for the two pixels within the region of interst. The function returns, the timeseries curves ploted together. The Id numbers can be obtained from the ASCII file outputs.
}
\examples{
TwoPointsPlot(114,125)


# The function results two time series vegetation index curves together at the plot pannel

}
\author{
Sofanit Araya
}
\seealso{
PhenoMetrics()
}
\keyword{Two}
\keyword{curves}
\keyword{point}
\keyword{time-series}

