% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(RawPath, BolAOI)
}
\arguments{
\item{Rawpath}{- Text value - the path where the time series images saved}

\item{BolAOI-}{Logical value - if there is any area of intererst or not}
}
\value{
Onset - the NDVI value at the start of the greenness

OnsetT - the time when the greeness of crop starts

MaxV - the annual maximum NDVI during Anthesis stage of crop

MaxT - the time when the maximum NDVI occurs

Offset - the NDVI value at the point before senesence stage of the crop

OffsetT - the time when the offset occured

GreenUpSlope- the rate of increase in NDVI between onset and Maximum NDVI

BrownDownSlope - The rate of decrease in NDVI from maximum NDVI to sensence

LengthGS- the length of the growing season between Onset and Offset

AreaBeforeMax - the integral area under the curve between Onset and Maximum NDVI

AreaAfterMax - the integral area under the curve between Maximum NDVI and Offser

TINDVI - the integral area under the curve

Asymmetry - the difference between AreaBeforeMax and AreaAfterMax
}
\description{
This function extracts major phenologic parameters from Moderate Resolution Imaging Spectroradiometer (MODIS)  time series vegetaion index data. Total of 11 phenologic metrics as raster and Ascii files. The function takes path of the vegetation index data and the boolean Value for BolAOI (True- if there is AOI polygon, FALSE- if the parameters are calculated for the whole region).
}
\details{
Remote sensing phenology also called Land Surface Phenology refers to observation of seasonal pattern of vegetation changes using remote sensing vegetation indices (Reed etal. 2009).
Remotely sensed vegetation phenology has been used for many ecological studies including as an indicator for climate change (Kramer et al.,2000; White et al., 1997), to estimate agricultural productivity (Hill and Donald, 2003; Labus et al., 2002; Sakamoto et al., 2013), regional management for crop type mapping (Brown et al., 2013; Niazmardi et al., 2013), as an indicator of soil Plant Available Water Holding Capacity (PAWC) variability across a farm (Araya etal. 2015)  and many more applications. 
Different methods have been employed to extract phenologic metrics, which include threshold definition (White et al., 1997), decomposition of the vegetation dynamic curve using harmonic analysis (Jakubauskas et al., 2001; Roerink et al., 2011) (Zhang et al., 2003) , taking the first derivative of the smoothed and non-smoothed vegetation index dynamics curves (Moulin et al., 1997) and defining the crossover point of the smoothed and non-smoothed dynamics curves (Hill and Donald, 2003; Reed et al., 1994). In this package the phenologic metrics were extracted based on correlation of the description of crop physiological stages (Zadoks etal. 1974) with the relative greenness of the crop on the vegetation dynamics. 
The list of the phenologic metrics and their description are provided at the website - www.cropphenology.wix.com/package

Detail of metrics definition

OnsetT and OnsetV
The OnsetT and OnsetV are defined as the value and time when the crop starts attaining high vegetation index increasingly. Thechnically, the algorthm used looks like as follows:

===================================================================================================

Assumption - 

1-Farm clearance done at – i.e. during time of  image 5 and image 6 so the trushold is considered to be mean of image 5 and image 6

trsh= (image5+image6)/2

2- The time frame for onset is between MODIS imaging time 7 and 12. We calculate the slope of the connecting line between values of images in that period

Slope=array of  (slope b/n 7 and 8, slope b/n 8 and 9, slope b/n 9 and 10, slope b/n 10 and 11, slope b/n 11 and 12

3- The slope observed next to trough is considered as Onset. So we calculate the last slope bellow - 0.01, just to avoid the minor details
                 
The last negative slope between image 7 and image 12
Case 1 - No negative slope (i.e all increasing)
 
 Step 1 - compare the values with the trsh value and the point where the trsh is excedded is Onset
    Case 1.1 - check if the next slop is –ve , although it is above -0.01
    
       If yes => take the next 
       
       If no => the point where the trsh is exceeded is Onset
       
    Case 1.2- if trsh is very high (sometimes due to uncleared weed or any other causes…)
    
                 Take the highest slope
                 
Case 2- last –ve is at the end of the onset time frame

 Check the previous slope (does it come from down or just a small trough?)
   
   Case 2.1 Came from down (previous is –ve)

             Check if the slope increases at image 12 then Onset is at 12

   Case 2.2 previous slope is +ve (it is a small trough)
   
             Ignore and consider it as all positive and take the point where threshold exceeded
Case3 - last -ve between 2 and 5
   
   Check previous slope
   
   Case 3.1 - if previous slope is +ve
         
         Consider it as a small trough and take the threshold exceeding point as Onset
         
   Case 3.2 - previous slope is -ve

        The onset is that point where the +ve slope started 

Case 4 - last -ve at the  beginning (i.e slope 1 or 2)

       Consider that as low vegetation after clearance and take the point where trsh exceeded as Onset
===================================================================================================

OffsetV and OffsetT

OffsetV and OffsetT are defined as the egetation index value and time when the plant senesence. On the timeseries vegetation curve, OffsetV and OffsetT are defined when the offset threshold value is attained
The threshold is defined as 10% above the total non green reference at the end pf the growing season, i.e average of 22 and 23 MODIS imaging periods.

MaxV and MaxT

MaxV and MaxT are defined as he value and time when the maximum vegetation index value attained during the growing season.

                                   Maximum (NDVI1, NDVI23)
}
\examples{
# EXAMPLE - 1

# phenologic metrics for region at the western Eyre Peninsula, South Australia. 
# The Raster images are clipped to the AOI, AOI = FALSE
PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE)


# EXAMPLE - 2
# Phenologic metrics for region at the Eastern Eyre Peninsula, South Australia. 
# Polygon boundary used as AOI from the MODIS images thus AOI= TRUE

PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE)

# In both examples the result will be saved at the working directory under the folder "Results"

}
\author{
Sofanit Araya
}
\references{
Araya, S., Lyle, G., Lewis, M., Ostendorf, B., in press. Phenologic metrics derived from MODIS NDVI as indicators for Plant Available Water-holding Capacity. Ecol. Ind.

Brown, J.C., Kastens, J.H., Coutinho, A.C., Victoria, D.d.C., Bishop, C.R., 2013. Classifying multiyear agricultural land use data from Mato Grosso using time-series MODIS vegetation index data. Rem. Sens. of Env/ 130, 39-50.

Hill, M.J., Donald, G.E., 2003. Estimating spatio-temporal patterns of agricultural productivity in fragmented landscapes using AVHRR NDVI time series. Rem.Sen. Env. 84, 367-384.

Moulin, S., Kergoat, L., Viovy, N., Dedieu, G., 1997. Global-scale assessment of vegetation phenology using NOAA/AVHRR satellite measurements. Journal of Climate 10, 1154-1170.

Reed, B.C., Schwartz, M.D., Xiao, X., 2009. Remote Sensing Phenology: Status and the Way Forward, in: Noormets, A. (Ed.), Phenology of Ecosystem Processes. Springer New York, pp. 231-246.

Roerink, G.J., Danes, M.H.G.I., Prieto, O.G., De Wit, A.J.W., Van Vliet, A.J.H., 2011. Deriving plant phenology from remote sensing, 2011 6th Int. Wor. on the, pp. 261-264.

Sakamoto, T., Gitelson, A.A., Arkebauer, T.J., 2013. MODIS-based corn grain yield estimation model incorporating crop phenology information. Rem. Sen.of Env. 131, 215-231.

White, M.A., Thornton, P.E., Running, S.W., 1997. A continental phenology model for monitoring vegetation responses to interannual climatic variability. Glo. Biogeochem. Cyc. 11, 217-234.

Zadoks, J.C., Chang, T.T., Konzak, C.F., 1974. A decimal code for the growth stages of cereals. Weed Research 14, 415-421.
}
\seealso{
MultiPointsPlot (N,Id1, Id2...Idn)
}
\keyword{Phenology,}
\keyword{image}
\keyword{remote}
\keyword{satellite}
\keyword{sensing,}

