% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(RawPath, BolAOI)
}
\arguments{
\item{Rawpath}{- Text value - the path where the time series images saved}

\item{BolAOI-}{Logical value - if there is any area of intererst or not}
}
\value{
Onset - the NDVI value at the start of the greenness

OnsetT - the time when the greeness of crop starts

MaxV - the annual maximum NDVI during Anthesis stage of crop

MaxT - the time when the maximum NDVI occurs

Offset - the NDVI value at the point before senesence stage of the crop

OffsetT - the time when the offset occured

GreenUpSlope- the rate of increase in NDVI between onset and Maximum NDVI

BrownDownSlope - The rate of decrease in NDVI from maximum NDVI to sensence

LengthGS- the length of the growing season between Onset and Offset

AreaBeforeMax - the integral area under the curve between Onset and Maximum NDVI

AreaAfterMax - the integral area under the curve between Maximum NDVI and Offser
}
\description{
Extract major phenologic parameters from time series vegetaion index data. Total of 11 phenologic metrics as raster and Ascii files. The function takes path of the vegetation index data and the boolean Value for BolAOI (True- if there is AOI polygon, FALSE- if the parameters are calculated for the whole region).
}
\examples{
PhenoMetrics("E:/MODIS/2005", TRUE)
}
\author{
Sofanit Araya
}
\references{
Araya etal.(2015)
}
\seealso{
TwoPointsPlot (Id1, Id2)
}
\keyword{Phenology,}
\keyword{image}
\keyword{remote}
\keyword{satellite}
\keyword{sensing,}

