% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackUp.R, R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(RawPath, BolAOI)

PhenoMetrics(RawPath, BolAOI)
}
\arguments{
\item{Rawpath}{- Text value - the path where the time series images saved}

\item{BolAOI-}{Logical value - if there is any area of intererst or not}

\item{Rawpath}{- Text value - the path where the time series images saved}

\item{BolAOI-}{Logical value - if there is any area of intererst or not}
}
\value{
Onset - the NDVI value at the start of the greenness

OnsetT - the time when the greeness of crop starts

MaxV - the annual maximum NDVI during Anthesis stage of crop

MaxT - the time when the maximum NDVI occurs

Offset - the NDVI value at the point before senesence stage of the crop

OffsetT - the time when the offset occured

GreenUpSlope- the rate of increase in NDVI between onset and Maximum NDVI

BrownDownSlope - The rate of decrease in NDVI from maximum NDVI to sensence

LengthGS- the length of the growing season between Onset and Offset

AreaBeforeMax - the integral area under the curve between Onset and Maximum NDVI

AreaAfterMax - the integral area under the curve between Maximum NDVI and Offser

TINDVI - the integral area under the curve

Asymmetry - the difference between AreaBeforeMax and AreaAfterMax

Onset - the NDVI value at the start of the greenness

OnsetT - the time when the greeness of crop starts

MaxV - the annual maximum NDVI during Anthesis stage of crop

MaxT - the time when the maximum NDVI occurs

Offset - the NDVI value at the point before senesence stage of the crop

OffsetT - the time when the offset occured

GreenUpSlope- the rate of increase in NDVI between onset and Maximum NDVI

BrownDownSlope - The rate of decrease in NDVI from maximum NDVI to sensence

LengthGS- the length of the growing season between Onset and Offset

AreaBeforeMax - the integral area under the curve between Onset and Maximum NDVI

AreaAfterMax - the integral area under the curve between Maximum NDVI and Offser

TINDVI - the integral area under the curve

Asymmetry - the difference between AreaBeforeMax and AreaAfterMax
}
\description{
This function extracts major phenologic parameters from time series vegetaion index data. Total of 11 phenologic metrics as raster and Ascii files. The function takes path of the vegetation index data and the boolean Value for BolAOI (True- if there is AOI polygon, FALSE- if the parameters are calculated for the whole region).

This function extracts major phenologic parameters from time series vegetaion index data. Total of 11 phenologic metrics as raster and Ascii files. The function takes path of the vegetation index data and the boolean Value for BolAOI (True- if there is AOI polygon, FALSE- if the parameters are calculated for the whole region).
}
\details{
Remote sensing phenology also called Land Surface Phenology refers to observation of seasonal pattern of vegetation changes using remote sensing vegetation indices (Reed etal. 2009).
Remotely sensed vegetation phenology has been used for many ecological studies including as an indicator for climate change (Kramer et al.,2000; White et al., 1997), to estimate agricultural productivity (Hill and Donald, 2003; Labus et al., 2002; Sakamoto et al., 2013), regional management for crop type mapping (Brown et al., 2013; Niazmardi et al., 2013), as an indicator of soil Plant Available Water Holding Capacity (PAWC) variability across a farm (Araya etal. 2015)  and many more applications. 
Different methods have been employed to extract phenologic metrics, which include threshold definition (White et al., 1997), decomposition of the vegetation dynamic curve using harmonic analysis (Jakubauskas et al., 2001; Roerink et al., 2011) (Zhang et al., 2003) , taking the first derivative of the smoothed and non-smoothed vegetation index dynamics curves (Moulin et al., 1997) and defining the crossover point of the smoothed and non-smoothed dynamics curves (Hill and Donald, 2003; Reed et al., 1994). In this package the phenologic metrics were extracted based on correlation of the description of crop physiological stages (Zadoks etal. 1974) with the relative greenness of the crop on the vegetation dynamics. 
The list of the phenologic metrics and their description are provided at the website - www.cropphenology.wix.com/package

Remote sensing phenology also called Land Surface Phenology refers to observation of seasonal pattern of vegetation changes using remote sensing vegetation indices (Reed etal. 2009).
Remotely sensed vegetation phenology has been used for many ecological studies including as an indicator for climate change (Kramer et al.,2000; White et al., 1997), to estimate agricultural productivity (Hill and Donald, 2003; Labus et al., 2002; Sakamoto et al., 2013), regional management for crop type mapping (Brown et al., 2013; Niazmardi et al., 2013), as an indicator of soil Plant Available Water Holding Capacity (PAWC) variability across a farm (Araya etal. 2015)  and many more applications. 
Different methods have been employed to extract phenologic metrics, which include threshold definition (White et al., 1997), decomposition of the vegetation dynamic curve using harmonic analysis (Jakubauskas et al., 2001; Roerink et al., 2011) (Zhang et al., 2003) , taking the first derivative of the smoothed and non-smoothed vegetation index dynamics curves (Moulin et al., 1997) and defining the crossover point of the smoothed and non-smoothed dynamics curves (Hill and Donald, 2003; Reed et al., 1994). In this package the phenologic metrics were extracted based on correlation of the description of crop physiological stages (Zadoks etal. 1974) with the relative greenness of the crop on the vegetation dynamics. 
The list of the phenologic metrics and their description are provided at the website - www.cropphenology.wix.com/package
}
\examples{
# EXAMPLE - 1

# phenologic metrics for region at the western Eyre Peninsula, South Australia. 
# The Raster images are clipped to the AOI, AOI = FALSE
PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE)


# EXAMPLE - 2
# Phenologic metrics for region at the Eastern Eyre Peninsula, South Australia. 
# Polygon boundary used as AOI from the MODIS images thus AOI= TRUE

PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE)

# In both examples the result will be saved at the working directory under the folder "Results"

# EXAMPLE - 1

# phenologic metrics for region at the western Eyre Peninsula, South Australia. 
# The Raster images are clipped to the AOI, AOI = FALSE
PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE)


# EXAMPLE - 2
# Phenologic metrics for region at the Eastern Eyre Peninsula, South Australia. 
# Polygon boundary used as AOI from the MODIS images thus AOI= TRUE

PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE)

# In both examples the result will be saved at the working directory under the folder "Results"

}
\author{
Sofanit Araya

Sofanit Araya
}
\references{
Araya, S., Lyle, G., Lewis, M., Ostendorf, B., in press. Phenologic metrics derived from MODIS NDVI as indicators for Plant Available Water-holding Capacity. Ecol. Ind.

Brown, J.C., Kastens, J.H., Coutinho, A.C., Victoria, D.d.C., Bishop, C.R., 2013. Classifying multiyear agricultural land use data from Mato Grosso using time-series MODIS vegetation index data. Rem. Sens. of Env/ 130, 39-50.

Hill, M.J., Donald, G.E., 2003. Estimating spatio-temporal patterns of agricultural productivity in fragmented landscapes using AVHRR NDVI time series. Rem.Sen. Env. 84, 367-384.

Moulin, S., Kergoat, L., Viovy, N., Dedieu, G., 1997. Global-scale assessment of vegetation phenology using NOAA/AVHRR satellite measurements. Journal of Climate 10, 1154-1170.

Reed, B.C., Schwartz, M.D., Xiao, X., 2009. Remote Sensing Phenology: Status and the Way Forward, in: Noormets, A. (Ed.), Phenology of Ecosystem Processes. Springer New York, pp. 231-246.

Roerink, G.J., Danes, M.H.G.I., Prieto, O.G., De Wit, A.J.W., Van Vliet, A.J.H., 2011. Deriving plant phenology from remote sensing, 2011 6th Int. Wor. on the, pp. 261-264.

Sakamoto, T., Gitelson, A.A., Arkebauer, T.J., 2013. MODIS-based corn grain yield estimation model incorporating crop phenology information. Rem. Sen.of Env. 131, 215-231.

White, M.A., Thornton, P.E., Running, S.W., 1997. A continental phenology model for monitoring vegetation responses to interannual climatic variability. Glo. Biogeochem. Cyc. 11, 217-234.

Zadoks, J.C., Chang, T.T., Konzak, C.F., 1974. A decimal code for the growth stages of cereals. Weed Research 14, 415-421.

Araya, S., Lyle, G., Lewis, M., Ostendorf, B., in press. Phenologic metrics derived from MODIS NDVI as indicators for Plant Available Water-holding Capacity. Ecol. Ind.

Brown, J.C., Kastens, J.H., Coutinho, A.C., Victoria, D.d.C., Bishop, C.R., 2013. Classifying multiyear agricultural land use data from Mato Grosso using time-series MODIS vegetation index data. Rem. Sens. of Env/ 130, 39-50.

Hill, M.J., Donald, G.E., 2003. Estimating spatio-temporal patterns of agricultural productivity in fragmented landscapes using AVHRR NDVI time series. Rem.Sen. Env. 84, 367-384.

Moulin, S., Kergoat, L., Viovy, N., Dedieu, G., 1997. Global-scale assessment of vegetation phenology using NOAA/AVHRR satellite measurements. Journal of Climate 10, 1154-1170.

Reed, B.C., Schwartz, M.D., Xiao, X., 2009. Remote Sensing Phenology: Status and the Way Forward, in: Noormets, A. (Ed.), Phenology of Ecosystem Processes. Springer New York, pp. 231-246.

Roerink, G.J., Danes, M.H.G.I., Prieto, O.G., De Wit, A.J.W., Van Vliet, A.J.H., 2011. Deriving plant phenology from remote sensing, 2011 6th Int. Wor. on the, pp. 261-264.

Sakamoto, T., Gitelson, A.A., Arkebauer, T.J., 2013. MODIS-based corn grain yield estimation model incorporating crop phenology information. Rem. Sen.of Env. 131, 215-231.

White, M.A., Thornton, P.E., Running, S.W., 1997. A continental phenology model for monitoring vegetation responses to interannual climatic variability. Glo. Biogeochem. Cyc. 11, 217-234.

Zadoks, J.C., Chang, T.T., Konzak, C.F., 1974. A decimal code for the growth stages of cereals. Weed Research 14, 415-421.
}
\seealso{
MultiPointsPlot (N,Id1, Id2...Idn)

MultiPointsPlot (N,Id1, Id2...Idn)
}
\keyword{Phenology,}
\keyword{image}
\keyword{remote}
\keyword{satellite}
\keyword{sensing,}

